Create README.md
Movie_API Project:

This project was made to create a movie database from scratch on the server-side using Node.js. 

Description:

This project includes APIs, web server frameworks, databases, business logic, authentication, and data security. In this database it will includes lists of movies, directors, genres, and users. It also allows for adding and removing of favorite movies. The tech stack used on this project is MERN (MongoDB, Express, React, and Node.js)